Generate RPC API documentation (#6)
This commit adds full documentation for all unit-e RPC calls. The
information is generated from the help information from the client
itself so it's easy to keep it up to date with the released versions
and with the latest development state. The commit represents the
current state of `dtr-org/unit-e` master.

It's done in a generic way so this could also be applied to upstream.
This might be a nice contribution to the documentation upstream.

There are still a few things missing and a few glitches:

* An intro to the RPC mechanism and how to use it is missing. Most of
  it can be taken from the upstream documentation.
* There are a couple of commands which have notes in the help which are
  different from how other commands are doing it. These notes are
  currently missing from the generated documentation.
* JSON structures are shown literally. They could also be parsed to
  allow better formatting.
* In general the formatting can be improved to make things clearer.

These things will be addressed in subsequent commits.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>