Add documentation of P2P messages (#21)
The documentation is generated from the list of messages from
the unit-e source code, the documentation which is part of the
source code, and detailed descriptions and definition of formats
from additional content currently stored in the directory
`helpers/p2p_docs/doc_data`. See `helpers/p2p_docs/README.md`
for more details.

The content of the documentation is based on the upstream
documentation from bitcoin.org and adapted to unit-e.

The script `helpers/p2p_docs/p2p_docs_helper` generates the RST
output stored in `reference/p2p` from which the HTML output is
generated by Sphinx.

The script also has a mode to check the examples in the documentation
and test them against the message classes defined in the Python
test framework in unit-e. The examples for messages from the
documentation are deserialized using the corresponding Python
classes from the unit-e test framework. The results are compared
with expected data.

The documentation covers the state which currently is in the unit-e
master branch.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>